Fix loxodromic midpoint on same circle of latitude
Currently, the Location.loxodromicMidPoint() method does not produce
correct results when the two locations have the same latitude. This is
because the intermediate value, lambda, is not finite in this case. For
points on the same circle of latitude, the midpoint's longitude may be
calculated by simply taking the mean of the input longitudes.