Cover additional edge cases in loxodromic midpoint
When both input locations lie along the same longitude, the
loxodromicMidPoint can be calculated in a simpler way, since the
midpoint's longitude will be the same as the input longitude.

This avoids issues primarily with the positive and negative antimeridian
lines at +180 and -180, which might otherwise return a midpoint with the
wrong sign after the trigonometric calculations take place.